backend/requirements.txt
Flask
pymongo
Flask-JWT-Extended
Flask-OAuthlib
boto3
python-dotenv

